+++ CHANGELOG.rst
@@ -1,2 +1,3 @@
-Next Release
-------------
+Initial release
+---------------
+- support for sending counters & timers to statsd over a TCP socket